Set up template engine and css framework.
Going to use mustache as the template engine we will be able to pass
values to it and place them within html. So mostly the templates will
be plain html but mustache allows iterating over lists and displaying
default values for things. I also added skeleton a CSS "framework"
it is just a really simple CSS file that sets up a grid system that
is responseive by default. Its that parts I wanted from bootstrap
without the large file size of bootstrap.