Added basic form for uploading images.
Includes a dropzone/multiple images allowed form element that takee
the chosen images and sends them to the server. The server then saves
the images in a tmp folder in the root directory (may need to be moved
depending on ability to serve those files directly back or could look
into saving them in local browser storage ...either way has to be stored
at least briefly on the server).

Once the image(s) has(ve) been uploaded the names of the files are added
to a div on the screen. This will eventually become a draggable list of
elements that will allow for ordering the images based on floor numbers.