Made the file names draggable/orderable.
Now the form page looks like a drop area/form for uploading images and a
box for the list of uploaded files.

After an image is uploaded its filename gets added to the list.
The list of file names is sortable/draggable. The next step is to
make another form that submits the image order to the server and
redirect to the page that will take inputs for start and end locations.