Added next screen of UI.
Basically this adds handling of the order of the draggable filenames and
translate them into form inputs to send back to the server. The server
then gets a lists of file names in order by floor. Onces it has this list
it redirects to a new page that displays the images in floor order.

This new page asks you to click two points on the images and displays a
blue dot on both places that you've clicked. The two points that you've
clicked are sent to the server ...and that's where this commit stops.

Up next is sending the point data and floor order to the path finding
scripts.