Now have a container for path information.
The container is called Directions (for now).
It has the start and end information and will eventually contain
everything there is to know about this path. This would make it easy to
eventually back this by a database or by a file in that you could just
serialize this class out and restore it. Its the state.

For now we just have this class and have added Gson dependency that will
turn this class into json and send it back to the calling javascript.

Refactored part of the matlab service to use this new Directions class.

Moved the css and js that I didn't write (ie third party/general library)
to a subfolder of the js and css folders to better differentiate for me
when I am looking for files to edit and also keeps apart all of the things
that would eventually need to be versioned or pulled from outside sources.