Save the points to exclude for matlab to use later
Also added a loading screen in case matlab takes a while to load.
(It does on my computer becuase matlab is installed on an external SD but
on a laptop that has matlab installed in a normal spot. The delay might
not be long at all.)

There was some merging error with DotGrid.m so I am commiting now so I can
look at that error separately.