Updated duplicate code in matlab (will change).
Deleted old _path.jpg images at the beginning of the new route requests.
Added error handling to parsing the path from matlab (it may still need to
do something different when no path is found but for some reason I ran
into a case where the last point in the path didn't have a y value so that
was what is being handled now). Added greying of buttons to try to
indicate which "mode" someone is in ie picking the two end points mode
versus picking points to exclude mode.